+++ Census/src/java/com/censo/controlador/documentos/ListarDocumentos.java
@@ -4,14 +4,12 @@ import com.censo.modelo.dao.CensoDao;
 import com.censo.modelo.dao.DocumentoDigitalizadoDao;
 import com.censo.modelo.persistencia.CenCenso;
 import com.censo.modelo.persistencia.CenDocumentosDigitalizado;
-import com.censo.modelo.persistencia.CenUsuario;
 import com.google.gson.Gson;
 import java.io.IOException;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.HashMap;
 import java.util.List;
-import java.util.LinkedList;
 import java.util.Map;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -57,7 +55,7 @@ public class ListarDocumentos extends HttpServlet {
             CensoDao censoDao = new CensoDao();
             conex = censoDao.conectar();
 
-            long idcenso = Long.parseLong(request.getParameter("idcenso"));
+            int idcenso = Integer.parseInt(request.getParameter("idcenso"));
             CenCenso cencenso = censoDao.ConsultarCensoById(conex, idcenso);
             if (cencenso == null) {
                 respuesta.put("status", "error");
