+++ Census/src/java/com/censo/controlador/persona/ModificarPersona.java
@@ -42,7 +42,7 @@ public class ModificarPersona extends HttpServlet {
                 return;
             }
 
-            long idpersona = Long.parseLong(request.getParameter("idpersona"));
+            int idpersona = Integer.parseInt(request.getParameter("idpersona"));
 
             PersonaDao personaDao = new PersonaDao();
             conex = personaDao.conectar();
@@ -145,7 +145,7 @@ public class ModificarPersona extends HttpServlet {
             }
             Date fechaNacimiento = new Date(new SimpleDateFormat("dd/MM/yyyy").parse(request.getParameter("txtfechanacimiento")).getTime());
             int genero = Integer.parseInt(request.getParameter("cmbgeneros"));
-            long municipio = Long.parseLong(request.getParameter("cmbmunicipios"));
+            int municipio = Integer.parseInt(request.getParameter("cmbmunicipios"));
             String direccion = request.getParameter("txtdireccion").toUpperCase().trim();
             String telefono = request.getParameter("txttelefono").toUpperCase().trim();
             String email = request.getParameter("txtemail").toUpperCase().trim();
