+++ Census/src/java/com/censo/controlador/usuario/ModificarUsuario.java
@@ -41,7 +41,7 @@ public class ModificarUsuario extends HttpServlet {
                 return;
             }
             
-            long idusuario = Long.parseLong(request.getParameter("idusuario"));
+            int idusuario = Integer.parseInt(request.getParameter("idusuario"));
 
             UsuarioDao usuarioDao = new UsuarioDao();
             conex = usuarioDao.conectar();
