+++ Census/src/java/com/censo/modelo/dao/DepartamentoDao.java
@@ -41,11 +41,11 @@ public class DepartamentoDao extends Conexion {
         return listaDepartamento;
     }
 
-    public CenDepartamento ConsultarDepartamentoById(Connection conex, long id) throws SQLException {
+    public CenDepartamento ConsultarDepartamentoById(Connection conex, int id) throws SQLException {
 
         try {
             pst = conex.prepareStatement("SELECT * FROM CEN_DEPARTAMENTOS WHERE DEPT_ID = ? ");
-            pst.setLong(1, id);
+            pst.setInt(1, id);
             rst = pst.executeQuery();
 
             while (rst.next()) {
